bb: rework terrain mirroring
Currently the 'bb_source' surface is used as a canvas to generate
chunks that are later cloned onto primary 'bb0' or 'bb1' surface -
formerly known as 'biter_battles'. The cloning step is very expensive
especially at the early stage of the game where chunks are forced to
generate to populate playing field.

This commit removes 'bb_source' surface completely and generates chunks
right on top of 'bbX' surface. Half of chunks are generated already for
north and no futher action is required to adjust them. This cuts amount
of operations game has to do by half which is visible in much faster
south side terrain mirroring and higher FPS.

During game over stage of the scenario we hot-swap players onto another
surface and the previous is deleted. This is necessary to workaround
engine limitation i.e. you cannot delete and create the same surface in
a single tick.

Since we turn some functions obsolete and rework core mechanics of the
scenario we also have to do changes in seemingly non-related areas like
biter's array of targets, silo generation - this unfortunately makes
commit bigger.