bb: terrain: fix "out-of-body" bug
Due to random chance after new surface is generated and player was
teleported to it when chunks around 0,0 are still generated - player's
character can be destroyed by accident. This happens in mixed ore
generation routine and is 100% reproducible with '1451460147' seed.

This bug was amplified by parent commit and will not reproduce with
given seed with previous version of this scenario. In my opinion this
is a data race.