bb: improve forced game reset logic
Check that we perform in biter_killed_the_silo relies on event 'cause'
field that can be absent according to documentation. This commit adds
another check based on force name on top that, so hopefully this is
sufficient as both of them are marked as optional.

We additionally remove biter spawning during silo death as it might die
before unit command is issued and cause crash. Also additional check is
added to see if force-map-reset command was already called.