bb: terrain: add support for configurable spawn area ores
Currently the spawn area ores parameters as well as distribution and
amount of patches is not configurable. Right now the logic is that
we count ores around starting area and draw the ones that do not appear
in starting area by accident.

This change adds a new system that first clears the starting area from
resources and then draws configurable amount of ore patches.