bb: terrain: fix truncated starting ores
Currently resources in starting position are cleared by searching a
fixed area. In case some ore vein ends up right on the edge of search
area it will cut in half and it will look not natural.
This patch makes the search area expand until no resource entity is
found or limit is reached.