terrain: optimize biter area texture generation
So far the method of generating tiles where biters spawn was very
resources intensive. There was a 2D simplex noise function ran 1024
times per each chunk in biter area. All of that was ran in single tick
which cause performance stutter. On top of that simplex noise generated
unrealistic pattern which some (me) find unappealing in the simplest form.

The proposed solution introduced a pre-computed seamless 2D texture
which we can map to x/y coordinates at virtually no cost. We only need
to compute offset in a texture grid, relative to position on map to
fetch tile name.