functions: boss_unit: filter out non-unit entities
Register on_entity_damaged event that is not triggered for every entity
that is being damanged. This is realized by applying filter. Also limit
the amount of sanity checks in the function itself. On logic, if entity
is valid, then it did not die yet and therefore it can be damaged.