terrain: river: reduce set_tiles calls
Right now, set_tiles is called per each individual tile in 32x32 chunk
when river is being generated. This is of course wasteful operation as
each time function is called, we need to switch context to C code,
apply single tile, correct the surrounding tiles and then come back to
lua script. This change buffers tiles and asks set_tiles to process
those in bulk once.

Fish don't care, those can be spawned in ground so we don't have to
be worried those are created before water is present. Unfortunately
entities cannot be spawned in bulk.