terrain: river: optimize memory use
At the moment each iteration cycle has to create position object and
reallocate tiles array for new element. A lot of time is wasted in
allocator. This change tricks the interpreter to pre-allocate entire
array by requiring it into scope with already predefined initlization
list. It can be local as we don't care about it's content, which is
safe to use in multiplayer environment.

This change also makes smarter use of position variable, only
allocating it when actually needed. Use of hash keys is also dropped
as those are not required by game engine.