perf : optim damage event  and increased refresh delay for healthbar of bosses (#400)
* functions: boss_unit: filter out non-unit entities

Register on_entity_damaged event that is not triggered for every entity
that is being damanged. This is realized by applying filter. Also limit
the amount of sanity checks in the function itself. On logic, if entity
is valid, then it did not die yet and therefore it can be damaged.

* functions: boss_unit: limit the amount of HP bar updates

At the moment the minimum grace period between health bar updates is
160ms. If the entity is damaged after that, few API calls will follow
to update state of HP bar. Limit that to once per half a second for
each boss for further optimization.