Refactor/standardize most code that prints out team names
This tries to standardize, within reason, how we access team names.

Background for this change:
 1. There are "force names" which are "north" and "south"
 2. There are the bb_config.(north|south)_side_team_name defaults
    "Team North" and "Team South"
 3. There are the configured global.tm_custom_name[] overridden
    values. Not overridden by default, but in captains games, these
    generally are.

Before this change, there was a combination of different things above
that were used when printing out a message, often with "team "
prepended. Specifically, it was common to print out ("team "
.. (tm_custom_name or force_name)) which would generally work
ok. However, the inconsistency was strange. Also, it was unclear when
choosing a custom team name if it should be "Team BuildMixers" or "The
BuildMixers" or just "BuildMixers", and it would sometimes be printed
out with "team " in front of it and sometimes not.

After this change, we try and relatively consistently use the new
Functions.team_name API to print out a team name, and do not prepend
it with "team " (because in the default case, it will be "Team North"
or "Team South"). A notable exception to this is when logging to
discord, we now consistently print out ("team " .. force_name),
because discord won't support the rich-text that is often used in
custom team names, and additionally bots may be parsing discord
messages and expect a specific structure..

Additionally, this adds team-name-coloring to the console output in a
few more places (like when a player joins a team, or when a technology
is researched).

For an added bonus, this change is net-negative in terms of lines of
code.

This change essentially leaves Discord logging unchanged, except in
one place (the "player.name has joined team X" the old code used
custom team names in Discord messages, and now it only uses "north"
and "south", which I suspect is much easier for parsers to parse (as
there is no other way for something parsing Discord messages to know
which team name is which team I think).