perf : Optimize river generation (#401)
* terrain: river: replace noise function with lookup array

It's expensive to run noise function per each tile in 4x32 grid (river
border). Replace that with pseudo random computed array with offset
values. Those values are in range of [-4,0] which is the same as old
design.

* terrain: river: reduce set_tiles calls

Right now, set_tiles is called per each individual tile in 32x32 chunk
when river is being generated. This is of course wasteful operation as
each time function is called, we need to switch context to C code,
apply single tile, correct the surrounding tiles and then come back to
lua script. This change buffers tiles and asks set_tiles to process
those in bulk once.

Fish don't care, those can be spawned in ground so we don't have to
be worried those are created before water is present. Unfortunately
entities cannot be spawned in bulk.

* terrain: river: optimize memory use

At the moment each iteration cycle has to create position object and
reallocate tiles array for new element. A lot of time is wasted in
allocator. This change tricks the interpreter to pre-allocate entire
array by requiring it into scope with already predefined initlization
list. It can be local as we don't care about it's content, which is
safe to use in multiplayer environment.

This change also makes smarter use of position variable, only
allocating it when actually needed. Use of hash keys is also dropped
as those are not required by game engine.

* terrain: move biter_texture to precomputed/ directory

As there are more and more machine generated files, move those to
dedicated directory.