precomputed: rework the biter_texture
In factorio 2.0 it's possible to pick up landfill, so let's get rid off
it from biter area. Because landfill was helping to shade nuclear-ground
the entire color palette needs to be changed as well to not look ugly.

Additionally, new tool appends header to the file so that texture can be
always reproduced.