functions.lua: make this file more "pure"
Specifically, with these changes (which are functionally a no-op), it
is now possible to load functions.lua outside of factorio (i.e. there
will be no access to gloabls/script/_STAGE/etc as part of loading the
file). This is being done to make it possible to write unit tests of
the functionality in functions.lua that run from the lua commandline
utiility.

This also marks a few more functions as local.