Speed up evo/threat feeding calculations
The calculations are no longer dependent on the number of
flasks. Specifically, each iteration of the loop is guaranteed to
either feed everything, or increase evo by at least 1% (for the range
of evo from 0% to 100%). After 100% evo, a single loop iteration will
finish it. Thus the loop will run at most ~100 times.

The new logic is essentially guaranteed to give very similar results
to the old logic. In the cases where one flask gives less than 1% evo
(which is by far by far the common case), this change is a very very
small buf to feeds (i.e. a feed that might go from 2% evo to 20% evo
will now go to 20.1% evo). In the case where one flask gives more than
1% evo (very rare), this change is a nerf to feeding. The test changes
demonstrate the (very small) delta in behavior.

My journey towards making this change was a long one. First I did a
lot of complex math to come up with an exact-approximation of the
current evo logic. However, doing the same thing for threat was going
to be even more complex, and the function was not invertible (or, it
was a very very complex invertion). When I gave up on that, I instead
decided to just keep the current logic, but turn it into a
piecewise-linear approximation, storing values every 5% evo. This
worked great, but was quite complex as it needed to maintain tables
for both evo and threat, and I would need to either change things to
store the total amount of food ever sent, or build functions to
convert the current evo value into the total amount of food that had
ever been sent. I made that all, and it worked, but it was a bunch of
code. When I was writing the initialization code (that would populate
the piecewise-linear table, I realized I could speed the code up by
not going one flask at a time, but by 1% evo at a time.  And finally,
I realized I could just do that from the get-go, and avoid a whole lot
of complexity, and thus here we are.  One advantage of the more
complex versions of the change is that the same amount of feed would
always result in the exact same evo values, i.e. the order of feed
would never make any difference. However, that doesn't seem terribly
important, and the difference from different orders should be quite
small.

This change is motivated by wanting to speed up the feeding code to no
longer be dependent on the number of flasks. This could enable
features like enabling users to see what the impact of a feed would be
without actually doing the feed.

--- here is the math/etc that can be ignored from my first attempt ---

We can take the existing code:
  local e2 = (clamped_evo * 100) + 1
  local diminishing_modifier = (1 / (10 ^ (e2 * 0.015))) / (e2 * 0.5)
  local evo_gain = (food_value * diminishing_modifier)

We can turn it into an exact/continuous function of the form
"total_food = f(evo)" by doing the following:

1. diminishing_modifier is the slope of the function that turns food
   into evo, so 1/diminishing_modifier is the slope of "f"
2. We can substitute in, getting
   slope = 1 / ((1 / (10 ^ (e2 * 0.015))) / (e2 * 0.5))
3. and then substitute in e2 as well getting
   slope = 1 / ((1 / (10 ^ ((evo * 100 + 1) * 0.015))) / ((evo * 100 + 1) * 0.5))
4. Due to dropping "clamped_evo" and just using "evo", the function
   will only be valid when evo <= 1.0
5. We can ask wolfram alpha [1] to integrate that slope, giving
   14.9852 e^(3.45388 x) (-0.27953 + x) + C
6. We can solve for C by knowing that total_food = 0, evo = 0 is a point
   14.9852 e^(3.45388 x) (-0.27953 + x) + 14.9852*0.27953
7. Note that the function after evo=1.0 is very simple, just completely linear.

And, by doing some spot checks, this function produces essentially
identical values as the current code. For instance, when giving green
sci, if we print out the f(evo)/actual_total_food_sent for each evo
value, we get a table like the following:

  evo:0.10 delta:1.002465
  evo:0.20 delta:1.000714
  evo:0.30 delta:1.000305
  evo:0.40 delta:1.000154
  evo:0.50 delta:1.000085
  evo:0.60 delta:1.000050
  evo:0.70 delta:1.000031
  evo:0.80 delta:1.000019
  evo:0.90 delta:1.000013
  evo:1.00 delta:1.000009
  evo:1.10 delta:1.083143
  evo:1.20 delta:1.293230

As expected, the function only stays valid up to evo:1.00, and after
that starts to diverge.

Note that if we send much-higher-tier science, the prediction is worse:
  evo:0.30 delta:8.439147
  evo:0.40 delta:1.653724
  evo:0.60 delta:1.121491
  evo:0.80 delta:1.037884
  evo:1.00 delta:1.013885

Note that it starts out at 30 because sending 1 flask of space science
jumps you all the way to ~28 evo with the current code (much much
higher than anyone would expect, because coming from 0 evo, the feed
is very very efficient. However, by the time we reach actually
significant evo values, things have just about evened out. I think
that this is a significant advantage of the new function
(specifically, giving the same food value worth of flasks will always
result in the same evo value), and I do not expect it to create an
substantial difference in real-world differences.

For threat calculations, we can theoretically do the same process, but
it is more complex, because it is in terms of evo, so I have not been
able to easily solve it.

[1] https://www.wolframalpha.com/input?i=integrate+1+%2F+%28%281+%2F+%2810+%5E+%28%28x+*+100+%2B+1%29+*+0.015%29%29%29+%2F+%28%28x+*+100+%2B+1%29+*+0.5%29%29

simplify