add /calc-send to see the effects of sending
By default, this will assume "to the other team" and "the contents of
the current player's inventory", but it is possible to directly
specify difficulty/evo/science values as well.

Note that I use `pcall` just in case there is some weird lua error
(for instance, an earlier version of this was missing an "if evo ==
nil" check, and so it could raise a Lua error about trying to do
arithmetic on a nil value, which would crash the scenario). If there
are weird arithmetic errors in the code, this should limit it to just
the specific command giving a strange error message.