Game start time: only start the time going up from 0 when something actually happens
This will change the "Time:" displayed in the GUI to display
"Not started" until the first player actually mines or crafts an item. I
think that this is especially valuable in captains games, which often
include 20+ minutes of team selection and discussion before the game
actually starts. I could potentially also hook into
on_player_changed_position to cover even more cases, but this seems
good enough for a first pass.

There are other ways that could potentially accomplish the goal behind
this change.  For instance, we could pause the timer whenever
global.freeze_players is set to true. This would work well for
captains games, but wouldn't help in public games where nobody joins
for 10+ minutes.

There is some slightly related logic in the passive feeding code (in
ai.lua:raise_evo) that doesn't do any passive feeding until after both
teams have players. However, I don't think that the logic is really
the same, and I don't think that it makes sense to tie these two.

There is potentially some interaction with the antigrief logic, which
will display raw game.ticks_played timestamps, which will no longer
agree with the "Time:" displayed in the UI.

Note that this change does not modify any of the difficulty voting
logic, so difficulty voting has to happen in the first 10 minutes from
the map load, which means that the GUI will no longer count down to
when difficulty votes are not allowed anymore.