Add a new tab displaying mutagen values and resource costs
I feel like this gives a lot of context for why sending higher-tier
sciences is worth more, and having this information available will
make new players learn the game faster. Note that I made some pretty
arbitrary decisions in the "cost" column. Specifically, which recipes
to assume the use of productivity modules for, and the relative cost
of each ore type and crude oil, as well as just blindly assuming that
all of the different petro/light/heavy oil are the same cost as
crude. That said, I think that this still gives a pretty accurate
picture of the costs/values of the different sciences.

Also, people have talked about potentially making a mode that slightly
randomizes the effectiveness of different sciences (or potentially
even randomizes what could be sent), and this tab would allow people
to decide what to throw in such a setting.

The function that I wrote to compute the raw resource cost of things
is slightly hacked together. Specifically, it won't be able to deal
with recipes that have multiple outputs very well (for instance, I
assume that it fails when given anything that requires uranium
processing). That said, I believe that it works for the current
science flask recipes fairly well (just, if you have lots of mods
enabled that add more complex recipes, then this will not work).

Many improvements/suggestions to this UI were suggested by people in
the community (i.e. carl/proxc both provided very helpful
suggestions).