Avoid "LuaTile API call when LuaTile was invalid." error
I get this error when I run the biter battles scenario with map
settings of width=50, height=50.  Not that this is a reasonable way to
run the scenario, but somehow this is how the settings are if I use
--load-scenario when starting factorio.  Also, even if this is not a
good choice/setting for running the game, it seems best if we don't
have any crashes/lua-errors when running in that configuration.