Remove legacy debug code (#439)
This is a partial PR from #410
The purpose of this is to simplify the codebase, and in doing so, make the gui-rewrite easier. By pushing this to a separate PR, it should be easier to review and evaluate. To my knowledge, none of this code is used by anyone.

I tested this code locally.
---------

Co-authored-by: DrButtons <DrButtons@empty.com>