Sync safe group-name with captain group picking
Group-names were previously colliding with the luaguielement parameters
and functions.  To overcome this, we started prepending '__' to each
group name using the comfy_panel.group helper function, see commit 0b53b2b

The captains code wasn't updated at that time, so this commit aims to
update the captains picking code to search for '__cpt' instead of 'cpt'
when selecting members of a group.

Additionally, the code wasn't updated to hide the __ when setting the
player's tag, which is visible from other players perspectives, as well
as from the 'map' perspective.  This commit aims to fix that as well.