Centralization of Event Handling Refactor 1
This code aims to centralize our event handling process in order to
improve code quality, and refactorability.

This specific commit:
- Centralizes on_player_mined_entity, on_player_mined_item,
  on_pre_player_crafted_item
- Adds type annotations to a small number of relevant functions

See discussion:

https://github.com/Factorio-Biter-Battles/Factorio-Biter-Battles/pull/440

Main post:

The current implementation is despised by the factorio devs. They have
been reluctant to help with fbb code in the past because they don't
think abstracting the on_event framework is smart for a variety of
reasons. There are multiple posts from them about this on the forums, I
think the spiciest is wrt to the stdlib library, but the message is
clear from them from all of those posts.
- Currently, when you have problems, the stacktrace doesn't go back any
  scopes, just up to Event.
- This is essentially a niche framework with complex requirements, for a
  niche set of code. I hope that by simplifying it to make it more like
  standard lua, and more like standard factorio code, we can encourage
  more contributors.
- Every event currently has its own set of if statements, we can combine
  them with single dispatch to hopefully improve performance.
- When debugging event problems, order is currently basically completely
  unknown. single dispatch makes it visible in the code.
- I had a minimal gui rewrite code setup, but I was having a ton of
  problems with the way that the events were being triggered in relation
  to eachother. I gave up eventually because every time I changed
  something, the events would happen out of order and buttons would move
  around. I don't remember the specifics, but it was frustrating.