Replace force-map-reset w/ instant-map-reset #364
The force-map-reset command randomly picks a silo and kills it in order
to trigger the map-reset timer.  This then requires you to wait the 120
seconds for the map reset, or - more often than not - set the gamespeed
to 5 and wait for the next map.

The instant-map-reset command essentially serves as a replacement to
that command, and has the benefits of:
- Not ending the game and sending mmr stats to the scores server.
- Is instant (resets the map immediately)
- Is reproducible (allows you to specify a seed)

Therefore, in the interest of de-bloating the codebase, it makes more
sense to keep just the instant-map-reset command, and remove the
force-map-reset command.