Deprecate utils/gui.lua handlers
Commented out:
- utils/gui.lua:Gui.on_elem_changed
- utils/gui.lua:Gui.on_selection_state_changed
- utils/gui.lua:Gui.on_pre_player_hide_top

We remove part of the gui.lua callbacks from being called as we are
moving all event handlers to the control.lua code.  This code only
serves to obfuscate what is actually happening, and is already barely
used.

With this I hope to make the gui-rewrite easier.

See discussion #440