Centralization of on_player_cursor_stack_changed
In preparation of the gui-rewrite, we move on_player_cursor_stack_changed
event handling from the antigrief.lua file to control.lua.
To do this we move the handler to a public function and call it from
control.lua.

The callback was modified to accept a validated LuaPlayer rather than
a raw event as passing the entire event was unnecessary.

See discussion at #440