Fix gui bugs Introduced during centralization
Part of the reason for the centralization of on-call events was to bring
visibility to the order of operations during   events like 'on_click'.
This manifested itself here, showing the importance of the order of
operations when constructing the gui and performing operations on it.

During the  refactor, order of operations was difficult to
diagnose (order of operations was decided by 'who gets required first').
This got the gui into a relatively working state by reordering and
adding ComfyPanelMain.on_gui_click which was somehow missed.

At this point in time the gui loads and appears to be working as
intended.  Further verification of integrity will be done at a later
date.

see discussion at #440