Add framework for rapid benchmarking
Statement:
We have long had problems with performance, and often have hacked
together crude benchmarks with various saves from public games.
However, because the lua state is serialized in a post-init state, it is
not easy to make dramatic changes to the code and evaluate them quickly.

Aim:
I hope that this PR will provide a simple, yet effective framework for
benchmarking changes to the code quickly.

Architecture:
While factorio has many options for benchmarking, it lacks the ability
to create a save file from a scenario through the commandline after a
specified number of ticks.  Further complicating this, game.server_save
only works on server instances, while --until-tick only works on single
player instances.  To get around this, we architect a workflow where we
use a bash script to manage chaining of multiple factorio commands, and
gate different parts of the script based on observed stdio, and tick #.

1. Preparation
1.a. Copy the working scenario directory to the factorio scenario dir
     under a new name.
1.b. Modify the copied scenario to enable the benchmarking code (by
     requiring 'benchmarking.main' in control.lua) via sed.

2. Initial save preparation
2.a. on_tick 100 set the map seed specifically, and then paste a
     defensive blueprint directly onto the new map.
2.b. Reveal related parts of the map to trigger chunk creation.
2.c. Turn on training mode and send a bunch of science to the biters in
     2 parts in hopes that the first batch has less boss biters to be a
     good benchmark, but still mildly realistic.
2.d. Wait until specific tick (determined via manual evaluation) for
     the biters to arrive.
2.e. Save the server to a specific name.
2.f. Once bash script detects stdout indicating the server has been
     saved, kill the server

3. Run benchmark
3.a. Start the factorio commandline benchmark from the previously made
     save for X number of ticks.
3.b. Use specific on_tick ~100 ticks after sever was saved, to enable
     the profiler, coordinating with X number of benchmark-ticks to
     ensure that the profiler exits before the tick limit.
3.c. Analyze the stdout via awk/grep, sorting reported calls by
     duration.

Required changes:
+ benchmarking folder to house these changes
+ benchmarking/server-settings.json file source:
  (https://github.com/wube/factorio-data/blob/master/server-settings.example.json)
  modified auto_pause to false so server starts without a player
  joining.
+ benchmarking/main.lua lua file that calls all lua-related code.
+ benchmarking/run_benchmark.sh bash script that coordinates the
  architecture.
+ commands/instant_map_reset.lua modified to enable calling from other
  lua files + formatting.
+ addition of a global: global.benchmark_mode so that biter groups
  can be created/selected even when no players are present.