Limit suspend voting to suspendee's force
Suspend was implemented but allowed people outside of the suspendee's
team to grief the vote by just randomly voting.  This limits the voter
pool to those on the suspendee's team.

To enable this, I had to add information about the suspender to the
global table, and to do this I wrapped most of the important
suspend-related information into a class 'SuspendTargetInfo'.  There are
parts of the global suspend table that could likely be included, but I
would like to keep this simple if possible.