Prevent using event framework for on_entity_damaged
I could also just change the single case that uses on_entity_damaged
to use the event framework, but we worry about the performance
overheads and are trying to move away from it in some cases anyways.

Anyhow, I am fine if we remove this in the future and just use the
event framework down the line.