New/faster/more-predictable revive mechanics (#472)
* benchmarks: small changes

Many of these are not clearly improvements, but they were useful
specifically for running benchmarks over the following changes
(changing biter revive mechanics). I also think that defaulting to
profiling=off makes more sense, because the only true performance
numbers are overall wallclock time. Looking at the profile can be
useful from time to time, but the reference is really the overall
walltime (which is drastically changed by the profiler being enabled).

* Improve benchmark reproducibility

Specifically, this means that much of the decisions about what biters
to spawn, and what biters will target, will be made independently of
any other calls to math.random() in the code. This is useful for
making a/b comparisons actually be apples-to-apples when using
benchmark.sh. Specifically, before this change, if you make a change
that ends up calling math.random during terrain generation, then it
would impact biter spawns/targets, and then potentially give
substantially different benchmark.sh results (because perhaps fewer
biters actually successfully attack, because they take a longer path
or something).

After this change, the biter behavior is identical even if other code
calls math.random() a different number of times.

* New biter revive logic

The high level idea behind this change is to change the way that
revive works. The helpful insight is that we can use
`on_entity_damaged` with a filter of `final_health <= 0` to identify
when biters would die, and we can "revive" them by setting them to
full health right there. This reduced a substantial amount of
complexity and improves performance.

The performance improvements come from (at least) two places:

1) We don't need to do extra work when entities die. Specifically, the
old code would look for corpses and destroy them/create new biters
where the corpses were. There was additional complexity from this
because the newly-spawned biters would not always behave the same as
the killed biters (i.e. might not have the same target/etc).  The new
units also potentially trigger new pathfinding (although this is just
a guess).

2) We no longer need to use `on_entity_damaged` to implement boss
biters.  We can just have a biter that will revive many many times for
a boss biter.

3) There is *maybe* a performance improvement from not
drawing/updating boss health bars, but it is unclear, and there might
be a performance regression from having more text-entities (displaying
the number of revives remaining), so this one is probably a wash.

I tried to make this change as little functionality as possible, this
is mostly just a performance change. There are some changes to
gameplay/graphics/etc.

1) Biters don't actually die/respawn. This means that the deaths don't
interrupt attack animations (so the biters are likely to attack
slightly more). Also, biter deaths don't reset slowdown-capsule
state. Also biter respawns aren't slightly delayed in time (the biter
stays alive). Also biters won't forget what orders they had when they
die/respawn.  This is likely a slight nerf to normal biters, due to
the slowdown capsule behavior. However, it seems like it is just
overall better playability.

2) Biter respawn chance is not nearly as random as it used to be.  It
used to be a completely random chance on each biter death.  So it was
at least theoretically possible that at 90% revive, one biter might
revive 40 times.  Now, at 90% revive, each biter will revive exactly 9
times.  At 33% revive, (randomly) half of the biters will revive
exactly 1 time, and the others will revive 0 times.  This is also
likely a very slight nerf to biters, because using poison to kill a
bunch of biters will result in them all dying at the same time
(instead of having a few stragglers where you get unlucky about the
revive count for just them).

3) Increasing evo% used to immediately impact all biters that
currently exist (giving them a higher revive chance). After this
change, the higher revive/health will only apply to newly spawned
biters.  Additionally, in the (not try_new_threat_logic) case, the
full threat-reduction for killing a high-health biter will be had for
killing any biter (including one spawned before evo% was high).  I
don't think that either of these will have a big impact in practice.
They make it slightly worse to send a bunch, wait 2 minutes, and then
send more. However, that is a bad strategy anyways.

4) Boss biters, when eventually killed, will essentially just die,
rather than potentially revive as normal biters.  At 90% revive, this
means that a boss essentially has 260x total health instead of 269x
total health (in expectation due to the additional expected 9
normal-biter revives).  This makes a pretty small difference in the
grand scheme of things.

I have made the code look fairly similar to the way that things looked
previously. The main change is that instead of bosses always showing
max-health, and having a healthbar that updates every ~1/2 second,
instead bosses will have health that goes down as normal, and their
health bar will shrink each time that their health is restored.

I think that there is substantial room for future improvements to
display the revives-remaining on both normal and boss biters in a
better/more-intuitive way.  If one runs
`/c global.bb_draw_revive_count_text = true`, then biters that are
spawned from then on will have text that shows the number of revives
remaining.  However, this text will be drawn on the ground (if it is
drawn above the layer that has biters, it is also drawn above the
map's fog-of-war, which seems unacceptable to me).

From my own runs of benchmark.sh, there is a ~36% overall performance
improvement (by subtracting out the time before the game actually
starts running) when lots of biters are attacking a well-defended
wall.

    % grep -e Goodbye -e 'consumed: -' /tmp/benchmark-???.log
    /tmp/benchmark-new.log:  12.177 Script ...
    /tmp/benchmark-new.log:  22.964 Goodbye
    /tmp/benchmark-old.log:  12.157 Script ...
    /tmp/benchmark-old.log:  26.854 Goodbye

* Revert lots of the math_random code changes.

They were useful for doing the benchmark, but don't really make the
code itself any better, and it is not very likely that we need this
level of benchmark-reproducibility for many more changes in the
future.

* Prevent using event framework for on_entity_damaged

I could also just change the single case that uses on_entity_damaged
to use the event framework, but we worry about the performance
overheads and are trying to move away from it in some cases anyways.

Anyhow, I am fine if we remove this in the future and just use the
event framework down the line.